Edit package hierarchy so classes are organized into Player, Game, and Universe.
