+++ android/app/src/main/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/entity/player/PlayerBuilder.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity;
+package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.player;
 
 /**
  * Builder pattern for the Player class.
