+++ android/app/src/main/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/entity/universe/Universe.java
@@ -1,0 +1,4 @@
+package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.universe;
+
+public class Universe {
+}
