+++ android/app/src/test/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/entity/ShipTester.java
@@ -1,7 +1,9 @@
 package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity;
 
-import org.junit.Before;
 import org.junit.Test;
+
+import edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.player.Ship;
+import edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.player.ShipType;
 
 import static org.junit.Assert.*;
